refactor: extract edit cv education section to own component
